moved browser middleware to browserMiddleware file, register middleware with name so that we can add middleware using before/after
